Add NeedNextTarget support
If we NeedNextTarget we now look forward in the queue for a
target as for MissingTarget. The difference is that if we find
a target in the queue we cancel the Failure and use that
target immediately in the entry (without propagating).

If there is no entry we trigger a normal failure and expect
the queue monitor to fix things up (with propagation).