Obtain projectid from MSB and not entry
For SCUBA-2 the non-science entries have their project ID stripped
so when the MSB is complete we can not rely on the entry for that
information so we get it from the Queue::MSB object instead.

We also now store the first non-CAL projectid in that object
when the MSB is added to the queue.