Display the warning messages from entry fixup.
When the fixup method returns warning messages, they are pushed
into an array in the entry object.  Then when the backend
is about to send the entry, it retrieves the warning messages
and passes them to its _pushmessage method.  This should result
in them being visible in the queue monitor.