Adjust setups to include observation track time.
When the fixup system determines the target for a SCUBA-2
setup, it should now add the tracking time of the following
observation to the setup's slew component.  This is to avoid
a setup being performed and then a large slew being needed
just afterwards, invalidating the setup.

See SciCom Trac ticket #140.