Add database timeouts
Specify a database timeout in ocsqmon and Queue::Server::DRAMA
before they connect to the database.  This should cause the
timeout to be configured when the connection is made (for
Sybase only in the current OMP code).  Timeouts are initially
set to 10 seconds.  This should avoid having the queue hang
if the database is being unresponsive.