Avoid accepting with blank shift type
Sometimes an MSB will be accepted with an entry which was not sent.
In that case its shift type will not have been set.  So now keep
searching until a non-blank (and non-undefined) value is found.
Also default to the current shift type, rather than "UNKNOWN",
if there is one, for the situation where we find no entries with
a shift type set.