Temporarily disable HTTPS host verification
Please see fault 20211004.001 -- we are currently unable to connect
to the OMP over HTTPS due to the expiry of the root certificate
underpinning our Let's Encrypt certificates.  It is not clear why
updating the Mozilla::CA package to remove the expired root
certificate only solved the issue on some machines and not others.
For more information about the issue, see also:
https://letsencrypt.org/2021/10/01/cert-chaining-help.html