Simplify layout of queue monitor window
This commit attempts to remove some of the unnecessary frames
from the queue monitor window in order to make the "info messages"
and "error messages" buttons work correctly.  However it seems
that the scrolling TextANSIColor widgets do not set their background
color when not inside a frame?

Increased the default height of the queue contents box and its
resizing weight in order to make it easier to increase the size
of this section of the window to see longer sequences of observations.