Allow catalog items of multiple source types
This updates the filtering function "source_is_type" to allow it
to operate on pointing catalog entries for which multiple types
are specified.  (No such entries exist at present, but this is to
allow it to be an option in future.) Added a test script to check
the behavior of this function.  Also clarified the documentation for
"source_is_type" now that there are more than two types.